+++ src/CollectSFDataDll/Common/Http.cs
@@ -19,7 +19,7 @@ namespace CollectSFData.Common
     {
         private readonly HttpClient _httpClient;
 
-        private static readonly CustomTaskManager _httpTasks = new CustomTaskManager();
+        private readonly CustomTaskManager _httpTasks = new CustomTaskManager();
 
         public bool DisplayError { get; set; }