+++ scripts/dotnet-build.ps1
@@ -3,8 +3,8 @@
 
 #>
 param(
-    [ValidateSet('net48', 'netcoreapp2.2', 'netcoreapp3.1', 'net7.0', 'net462')]
-    [string[]]$targetFrameworks = @('net48', 'netcoreapp3.1', 'net7.0', 'net462'),
+    [ValidateSet('net48', 'netcoreapp2.2', 'netcoreapp3.1', 'net6.0', 'net7.0', 'net462')]
+    [string[]]$targetFrameworks = @('net48', 'netcoreapp3.1', 'net6.0', 'net7.0', 'net462'),
     [ValidateSet('all', 'debug', 'release')]
     $configuration = 'all',
     [switch]$publish,